% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GISsource.r
\name{AltXY7}
\alias{AltXY7}
\title{read DEM .txt data of JGD7
 @param DEMdir directory of DEM data}
\usage{
AltXY7(DEMdir, JGD7dv, JGD7dh)
}
\arguments{
\item{JGD7dv}{vertical number}

\item{JGD7dh}{horizontal number}
}
\value{
dem data
}
\description{
read DEM .txt data of JGD7
 @param DEMdir directory of DEM data
}
\examples{
# not run
}
