% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GISsource.r
\name{area}
\alias{area}
\title{calculate polygonal area}
\usage{
area(xl, yl)
}
\arguments{
\item{xl}{x coordinates}

\item{yl}{y coordinates}
}
\value{
numeric
}
\description{
calculate polygonal area
}
\examples{
xl <- c(0,1,2,1)
yl <- c(1,2,1,0)
plot(xl,yl,type="b") ; polygon(xl,yl)
area(xl,yl)

}
