% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster2.R
\name{dchm_raster}
\alias{dchm_raster}
\title{Return a raster of Digital Crown Height Model(DCHM) from DSM x,y,z, and DTM raster}
\usage{
dchm_raster(x = dsm$x, y = dsm$y, dtm_raster = r)
}
\arguments{
\item{x}{DSM x}

\item{y}{DSM y}

\item{r}{DTM raster}
}
\value{
a raster of Digital Crown Height Model(DCHM)
}
\description{
Return a raster of Digital Crown Height Model(DCHM) from DSM x,y,z, and DTM raster
}
\examples{
# load ortho GeoTif ####
.dir<-system.file("extdata",package="raster2")
filename <- paste0(.dir,"/07KE0024_office.tif")
 b<-raster::brick(filename)
raster::plotRGB(b)
# 3D graphics ####
v<-raster::getValues(b)
rgb_<-v[cellFromXY(b, dsm[,c("x","y")]),]
rgl::open3d()
rgl::points3d(dsm,col=rgb(rgb_,maxColorValue = 255))
filename <- paste0(.dir,"/dtm.tif")
dtm_raster<-raster::raster(filename)
contour(dtm_raster)
CHM<-dchm_raster(dsm$x,y=dsm$y,dtm_raster)
plot(CHM)
# ForestTools::vwf
lin <- function(x){x * 0.05 + 0.6}
ttops <- ForestTools::vwf(CHM = CHM, winFun = lin, minHeight = 2)
plot(ttops, col = "blue", pch = 20, cex = 0.5, add = TRUE) # plot of ttops
# Plot crowns
crowns <- ForestTools::mcws(treetops = ttops, CHM = CHM, minHeight = 1.5)
plot(crowns, col = sample(rainbow(50), length(unique(crowns[])), replace = TRUE), legend = FALSE, xlab = "", ylab = "", xaxt='n', yaxt = 'n')

tt<-data.frame(ttops,sf::st_coordinates(ttops)) #write.csv(tt,file="Kuraiyama_Office_ttop.csv")
head(tt)
#### 地盤高データ追加

Z<-raster::getValues(dtm_raster)[cellFromXY(dtm_raster,tt[,c("X","Y")])]
tt<-data.frame(tt,Z)
head(tt)
####　地盤高水平化
dchm <- dchm_vector(x=dsm$x,y=dsm$y,dtm_raster)
rgl::open3d()
rgl::points3d(dchm[dchm$h <0.5,],col="red")
rgl::points3d(dchm[dchm$h >= 0.5 & dchm$h <2 ,],col="green")
rgl::points3d(dchm,col=rgb(rgb_,maxColorValue = 255))
n<-nrow(tt)
i<-rep(seq(n),each=2)
trunk_xyz<-tt[i,c("X","Y","height")]
trunk_xyz$height[2*seq(n)]<-0
rgl::segments3d(trunk_xyz)

#####　地盤高　+ dsm + オルソ　+　立木

rgl::open3d()
h<-dchm$h
rgl::points3d(dsm[h >= 0.5 & h < 2 ,],col="green")
rgl::points3d(dsm[h <0.5,],col="magenta")
rgl::points3d(dsm,col=rgb(rgb_,maxColorValue = 255))
n<-nrow(tt)
i<-rep(seq(n),each=2)
trunk_xyz_dsm<-tt[i,c("X","Y","Z")]
trunk_xyz_dsm$Z[2*seq(n)]<-tt$Z+tt$height
rgl::segments3d(trunk_xyz_dsm)
rgl::points3d(tt$X,tt$Y,tt$Z+tt$height,col="red",size=10)
}
