% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster2.R
\name{dchm_vector}
\alias{dchm_vector}
\title{Return a vector of Digital Crown Height Model(DCHM) from DSM x,y,z, and DTM raster}
\usage{
dchm_vector(x = dsm$x, y = dsm$y, dtm_raster = r)
}
\arguments{
\item{x}{DSM x}

\item{y}{DSM y}

\item{r}{DTM raster}
}
\value{
a vector of Digital Crown Height Model(DCHM)
}
\description{
Return a vector of Digital Crown Height Model(DCHM) from DSM x,y,z, and DTM raster
}
\examples{
# load ortho GeoTif ####
.dir<-system.file("extdata",package="raster2")
filename <- paste0(.dir,"/07KE0024_office.tif")
 b<-raster::brick(filename)
raster::plotRGB(b)
# 3D graphics ####
v<-raster::getValues(b)
rgb_<-v[cellFromXY(b, dsm[,c("x","y")]),]
rgl::open3d()
rgl::points3d(dsm,col=rgb(rgb_,maxColorValue = 255))
# DTM
filename <- paste0(.dir,"/dtm.tif")
dtm_raster<-raster::raster(filename)
contour(dtm_raster)
# dchm_vector
dchm <- dchm_vector(x=dsm$x,y=dsm$y,dtm_raster)
rgl::open3d()
rgl::points3d(dchm,col=rgb(rgb_,maxColorValue = 255))
}
