% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GISsource.r
\name{Grass2R_AsciiRaster}
\alias{Grass2R_AsciiRaster}
\title{Grass ASCII raster grid -> R image}
\usage{
Grass2R_AsciiRaster(filename)
}
\arguments{
\item{filename}{}
}
\description{
Grass ASCII raster grid -> R image
}
