% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GISsource.r
\name{l2r}
\alias{l2r}
\title{Title}
\usage{
l2r(xaxis, yaxis, px00, py00)
}
\arguments{
\item{py00}{}
}
\description{
Title
}
\examples{


xaxis <-seq(0,21,.1) ;yaxis <-seq(0,21,.1)
px00<-c(1,10,21,10,1) ;  py00<-c(10,20,10,1,10)
px00<-c(1,10) ;  py00<-c(10,20)
r<-l2r(xaxis,yaxis,px00,py00)
image(r)
}
