% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GISsource.r
\name{R2Grass_AsciiRaster}
\alias{R2Grass_AsciiRaster}
\title{R image ->   Grass ASCII raster grid}
\usage{
R2Grass_AsciiRaster(we, sn, m, filename)
}
\arguments{
\item{filename}{}
}
\description{
R image ->   Grass ASCII raster grid
}
