% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GISsource.r
\name{radian}
\alias{radian}
\title{convert to radian from degree}
\usage{
radian(degree)
}
\arguments{
\item{degree}{}
}
\value{
radian value
}
\description{
convert to radian from degree
}
\examples{
radian(360)/pi
}
