% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster2.R
\name{ras2img}
\alias{ras2img}
\title{ras2img}
\usage{
ras2img(m)
}
\arguments{
\item{m}{"raster" matrix}
}
\value{
"image" matrix
}
\description{
convert from "raster" matrix to "image" matrix
}
\examples{
(img<-matrix(1:9,3))
img2ras(img)
ras2img(img2ras(img))


}
