% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GISsource.r
\name{s2n}
\alias{s2n}
\title{角度表現　度分秒<->度 (カシミール3d書式)}
\usage{
s2n(d)
}
\arguments{
\item{d}{}
}
\value{
numeric degree
}
\description{
角度表現　度分秒<->度 (カシミール3d書式)
}
\examples{
s2n("36.300000")
s2n("36.5959900")

}
