% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster2.R
\name{surface3dr}
\alias{surface3dr}
\title{surface3dr
draw surface 3d from raster data}
\usage{
surface3dr(r, ...)
}
\arguments{
\item{r}{}
}
\value{
No return values, only draw surface3d.
}
\description{
surface3dr
draw surface 3d from raster data
}
\examples{
.dir<-system.file("extdata",package="GISsource")
filename <- paste0(.dir,"/dtm.tif")
r<-raster::raster(filename)　　　
plot(r)
surface3dr(r,col="gray")

}
