% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GISsource.r
\name{TotalLength}
\alias{TotalLength}
\title{Total length of circumference of polygon XY}
\usage{
TotalLength(XY)
}
\arguments{
\item{XY}{XY matrix}
}
\value{
Total length
}
\description{
Total length of circumference of polygon XY
}
\examples{
XY <- cbind(X=c(1,2,3,2,1),Y=c(1,2,1,0,1))
plot(XY,type="b")
TotalLength(XY)
}
