% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GISsource.r
\name{wpt}
\alias{wpt}
\title{wptXY　カシミールのウェイポイント・データからXY、ラベルを抽出する関数}
\usage{
wpt(filename)
}
\arguments{
\item{filename}{}
}
\value{
x,y coordinates of way points (.wpt)
}
\description{
wptXY　カシミールのウェイポイント・データからXY、ラベルを抽出する関数
}
\examples{
 .dir<-system.file("extdata",package="GISsource")
  filename <- paste0(.dir,"/演習林範囲.wpt")
  wpt(filename)

カシミールwptファイルの行からラベルデータ取得　
}
