% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster2.R
\name{xy2ij}
\alias{xy2ij}
\title{Return  reference number i,j of raster matrix from x, y coordinates}
\usage{
xy2ij(x, y, r)
}
\arguments{
\item{x}{coordinates}

\item{y}{coordinates}

\item{r}{class of raster data}
}
\value{
integer matrix of reference number i,j
}
\description{
Return  reference number i,j of raster matrix from x, y coordinates
}
\examples{
# load DTM data
filename <- paste0(.dir,"/dtm.tif")
dtm_raster<-raster::raster(filename)
plot(dtm_raster)

(ij <- xy2ij(3900:3910,-1100:-1110,dtm_raster))

dtm_raster[ij]   # get altitude for x,y from raster matrix

}
