% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GISsource.r
\name{xy2ll}
\alias{xy2ll}
\title{Title}
\usage{
xy2ll(x, y, lat00 = "36.00000", lon00 = "137.100000")
}
\arguments{
\item{lon00}{}
}
\value{
latitudde and longitude
}
\description{
Title
}
\examples{
xy2ll(0,0)
}
