% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster2.R
\name{xy2n}
\alias{xy2n}
\title{Return cell number of raster matrix from x,y coordinates}
\usage{
xy2n(x, y, r)
}
\arguments{
\item{x}{coordinates}

\item{y}{coordinates}

\item{r}{class of raster data}
}
\value{
integer of cell number
}
\description{
Return cell number of raster matrix from x,y coordinates
}
\examples{

xy2n(x,y,r)

}
