% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GISsource.r
\name{XY7Code}
\alias{XY7Code}
\title{平面直角座標系第7系 Map Code}
\usage{
XY7Code(JGD7dv, JGD7dh)
}
\arguments{
\item{JGD7dv}{vertical number}

\item{JGD7dh}{horizontal number}
}
\value{
list of map information
}
\description{
平面直角座標系第7系 Map Code
}
\examples{
XY7Code(1,1)

}
