% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster2.R
\name{xyz_ext}
\alias{xyz_ext}
\title{crop xyz data matrix or data frame form extent c(x1,x2,y1,y2)}
\usage{
xyz_ext(xyz = dtm, ext = c(5100, 5200, -1400, 1100))
}
\arguments{
\item{xyz}{matrix of x,y,z coordinates}

\item{ext}{matrix of x,y,z coordinates}
}
\value{
xyz data matrix or data frame
}
\description{
crop xyz data matrix or data frame form extent c(x1,x2,y1,y2)
}
\examples{
ext <-c(3850,3900,-1100,-1050)
xyz_ext(dtm,ext)
}
