% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GISsource.r
\name{n2s}
\alias{n2s}
\title{Title}
\usage{
n2s(d)
}
\arguments{
\item{d}{}
}
\value{

}
\description{
Title
}
\examples{
d=137.13416566
n2s(d)
s2n(n2s(d))
#########  第7系原点
lat0 = s2n("36.00000")*degree;
lon0 = s2n("137.100000")*degree;
}
